move fanoutconsumer from service/internal to internal
