+++ internal/fanoutconsumer/metrics_test.go
